Create architecture
Milestone 2 of the DEVxDAO Grant 305 Garment Chain.  This describes the architecture of how a Physically-Based Token system could be combined with one using composable NFTs to show full transparency and traceability of garment supply chains.  Specific examples of how this system would work are explained, demonstrating in which scenarios it would work well, and in which scenarios it may have less added value.  The same concept could also be applied to other sectors.